Add hook for customizers
